@@ -4,13 +4,15 @@ line_length = 120
 
 [tool.black]
 line-length = 120
-target-version = ['py38']
+target-version = ['py310']
 
 [tool.ruff]
 line-length = 120
-target-version = "py38"
+
+[tool.ruff.lint]
 select = ["E", "F", "W", "I"]
+fixable = ["ALL"]
 ignore = []
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"] # Ignore unused imports in __init__.py files
